requirements.txt
fastapi
uvicorn
PyPDF2
langchain
langchain-community
langchain-core
langchain-openai
spacy
faiss-cpu
python-dotenv
python-multipart
streamlit 
requests
